reorganize the tests to follow keras best practices
