+++ keras_tuner/engine/base_tuner_test.py
@@ -16,7 +16,6 @@ import os
 import pickle
 
 import numpy as np
-import pytest
 from sklearn import linear_model
 
 import keras_tuner
@@ -31,12 +30,7 @@ VAL_INPUTS = np.random.random(size=(NUM_SAMPLES, INPUT_DIM))
 VAL_TARGETS = np.random.randint(0, NUM_CLASSES, size=(NUM_SAMPLES,))
 
 
-@pytest.fixture(scope="function")
-def tmp_dir(tmpdir_factory):
-    return tmpdir_factory.mktemp("integration_test", numbered=True)
-
-
-def test_base_tuner(tmp_dir):
+def test_base_tuner(tmp_path):
     class MyTuner(base_tuner.BaseTuner):
         def run_trial(self, trial, x):
             model = self.hypermodel.build(trial.hyperparameters)
@@ -62,7 +56,7 @@ def test_base_tuner(tmp_dir):
     oracle = keras_tuner.tuners.randomsearch.RandomSearchOracle(
         objective=keras_tuner.Objective("score", "max"), max_trials=5
     )
-    tuner = MyTuner(oracle=oracle, hypermodel=build_model, directory=tmp_dir)
+    tuner = MyTuner(oracle=oracle, hypermodel=build_model, directory=tmp_path)
     tuner.search(1.0)
     models = tuner.get_best_models(5)
 
@@ -71,7 +65,7 @@ def test_base_tuner(tmp_dir):
     assert models[0] == models_by_factor[0]
 
 
-def test_simple_sklearn_tuner(tmp_dir):
+def test_simple_sklearn_tuner(tmp_path):
     class SimpleSklearnTuner(base_tuner.BaseTuner):
         def run_trial(self, trial, x, y, validation_data):
             model = self.hypermodel.build(trial.hyperparameters)
@@ -100,7 +94,7 @@ def test_simple_sklearn_tuner(tmp_dir):
             objective=keras_tuner.Objective("score", "max"), max_trials=2
         ),
         hypermodel=sklearn_build_fn,
-        directory=tmp_dir,
+        directory=tmp_path,
     )
     tuner.search(
         TRAIN_INPUTS, TRAIN_TARGETS, validation_data=(VAL_INPUTS, VAL_TARGETS)
