+++ keras_tuner/engine/tuner_test.py
@@ -19,10 +19,12 @@ from unittest.mock import patch
 import numpy as np
 import pytest
 import tensorflow as tf
+from packaging.version import parse
 from tensorboard.plugins.hparams import api as hparams_api
 from tensorflow import keras
 
 import keras_tuner
+from keras_tuner.engine import tuner as tuner_module
 
 INPUT_DIM = 2
 NUM_CLASSES = 3
@@ -31,11 +33,6 @@ TRAIN_INPUTS = np.random.random(size=(NUM_SAMPLES, INPUT_DIM))
 TRAIN_TARGETS = np.random.randint(0, NUM_CLASSES, size=(NUM_SAMPLES, 1))
 VAL_INPUTS = np.random.random(size=(NUM_SAMPLES, INPUT_DIM))
 VAL_TARGETS = np.random.randint(0, NUM_CLASSES, size=(NUM_SAMPLES, 1))
-
-
-@pytest.fixture(scope="function")
-def tmp_dir(tmpdir_factory):
-    return tmpdir_factory.mktemp("integration_test", numbered=True)
 
 
 def build_model(hp):
@@ -55,6 +52,75 @@ def build_model(hp):
         metrics=["accuracy"],
     )
     return model
+
+
+class MockModel(keras.Model):
+    def __init__(self, full_history):
+        super(MockModel, self).__init__()
+        self.full_history = full_history
+        self.callbacks = []
+        self.optimizer = True
+
+    def call_callbacks(self, callbacks, method_name, *args, **kwargs):
+        for callback in callbacks:
+            method = getattr(callback, method_name)
+            method(*args, **kwargs)
+
+    def on_epoch_begin(self, epoch):
+        for callback in self.callbacks:
+            callback.on_epoch_begin(epoch, logs=None)
+
+    def on_epoch_end(self, epoch):
+        logs = {"loss": np.average(self.full_history[epoch])}
+        for callback in self.callbacks:
+            callback.on_epoch_end(epoch, logs=logs)
+
+    def on_batch_begin(self, epoch, batch):
+        for callback in self.callbacks:
+            callback.on_batch_begin(batch, logs=None)
+
+    def on_batch_end(self, epoch, batch):
+        logs = {"loss": self.full_history[epoch][batch]}
+        for callback in self.callbacks:
+            callback.on_batch_end(epoch, logs=logs)
+
+    def fit(self, *args, **kwargs):
+        self.callbacks = kwargs["callbacks"]
+        for callback in self.callbacks:
+            callback.model = self
+        for epoch in range(len(self.full_history)):
+            self.on_epoch_begin(epoch)
+            for batch in range(len(self.full_history[epoch])):
+                self.on_batch_begin(epoch, batch)
+                self.on_batch_end(epoch, batch)
+            self.on_epoch_end(epoch)
+        history = keras.callbacks.History()
+        history.history = {
+            "loss": [np.average(epoch_values) for epoch_values in self.full_history]
+        }
+        return history
+
+    def save_weights(self, fname, **kwargs):
+        pass
+
+    def get_config(self):
+        return {}
+
+
+class MockHyperModel(keras_tuner.HyperModel):
+
+    mode_0 = [[10, 9, 8], [7, 6, 5], [4, 3, 2]]
+    mode_1 = [[13, 13, 13], [12, 12, 12], [11, 11, 11]]
+
+    def __init__(self):
+        # The first call to `build` in tuner __init__
+        # will reset this to 0
+        self.mode_0_execution_count = -1
+
+    def build(self, hp):
+        if hp.Choice("mode", [0, 1]) == 0:
+            return MockModel(self.mode_0)
+        return MockModel(self.mode_1)
 
 
 def build_subclass_model(hp):
@@ -109,19 +175,19 @@ class ExampleHyperModel(keras_tuner.HyperModel):
         return model.fit(*args, shuffle=hp.Boolean("shuffle"), **kwargs)
 
 
-def test_basic_tuner_attributes(tmp_dir):
+def test_basic_tuner_attributes(tmp_path):
     tuner = keras_tuner.tuners.RandomSearch(
         build_model,
         objective="val_accuracy",
         max_trials=2,
         executions_per_trial=3,
-        directory=tmp_dir,
+        directory=tmp_path,
     )
 
     assert tuner.oracle.objective.name == "val_accuracy"
     assert tuner.oracle.max_trials == 2
     assert tuner.executions_per_trial == 3
-    assert tuner.directory == tmp_dir
+    assert tuner.directory == tmp_path
     assert tuner.hypermodel.__class__.__name__ == "DefaultHyperModel"
     assert len(tuner.oracle.hyperparameters.space) == 3  # default search space
     assert len(tuner.oracle.hyperparameters.values) == 3  # default search space
@@ -138,16 +204,16 @@ def test_basic_tuner_attributes(tmp_dir):
     tuner.results_summary()
 
     assert len(tuner.oracle.trials) == 2
-    assert os.path.exists(os.path.join(str(tmp_dir), "untitled_project"))
+    assert os.path.exists(os.path.join(str(tmp_path), "untitled_project"))
 
 
-def test_multi_objective(tmp_dir):
+def test_multi_objective(tmp_path):
     tuner = keras_tuner.tuners.RandomSearch(
         build_model,
         objective=["val_accuracy", "val_loss"],
         max_trials=2,
         executions_per_trial=3,
-        directory=tmp_dir,
+        directory=tmp_path,
     )
 
     assert tuner.oracle.objective.name == "multi_objective"
@@ -164,7 +230,7 @@ def test_multi_objective(tmp_dir):
     tuner.results_summary()
 
 
-def test_no_hypermodel_with_objective(tmp_dir):
+def test_no_hypermodel_with_objective(tmp_path):
     class MyTuner(keras_tuner.tuners.RandomSearch):
         def run_trial(self, trial, *args, **kwargs):
             hp = trial.hyperparameters
@@ -173,14 +239,14 @@ def test_no_hypermodel_with_objective(tmp_dir):
     tuner = MyTuner(
         objective="val_loss",
         max_trials=2,
-        directory=tmp_dir,
+        directory=tmp_path,
     )
     tuner.search()
 
     assert len(tuner.oracle.trials) == 2
 
 
-def test_no_objective_with_hypermodel(tmp_dir):
+def test_no_objective_with_hypermodel(tmp_path):
     class MyHyperModel(ExampleHyperModel):
         def fit(self, hp, model, *args, **kwargs):
             return hp.Float("value", 0, 10)
@@ -189,14 +255,14 @@ def test_no_objective_with_hypermodel(tmp_dir):
         hypermodel=MyHyperModel(),
         max_trials=2,
         executions_per_trial=3,
-        directory=tmp_dir,
+        directory=tmp_path,
     )
     tuner.search()
 
     assert len(tuner.oracle.trials) == 2
 
 
-def test_no_hypermodel_no_objective(tmp_dir):
+def test_no_hypermodel_no_objective(tmp_path):
     class MyTuner(keras_tuner.tuners.RandomSearch):
         def run_trial(self, trial, *args, **kwargs):
             hp = trial.hyperparameters
@@ -205,23 +271,23 @@ def test_no_hypermodel_no_objective(tmp_dir):
     tuner = MyTuner(
         objective="val_loss",
         max_trials=2,
-        directory=tmp_dir,
+        directory=tmp_path,
     )
     tuner.search()
 
     assert len(tuner.oracle.trials) == 2
 
 
-def test_no_hypermodel_without_override_run_trial_error(tmp_dir):
+def test_no_hypermodel_without_override_run_trial_error(tmp_path):
     with pytest.raises(ValueError, match="Received `hypermodel=None`"):
         keras_tuner.tuners.RandomSearch(
             max_trials=2,
             executions_per_trial=3,
-            directory=tmp_dir,
+            directory=tmp_path,
         )
 
 
-def test_no_objective_return_not_single_value_error(tmp_dir):
+def test_no_objective_return_not_single_value_error(tmp_path):
     class MyHyperModel(ExampleHyperModel):
         def fit(self, hp, model, *args, **kwargs):
             return {"val_loss": hp.Float("value", 0, 10)}
@@ -230,20 +296,20 @@ def test_no_objective_return_not_single_value_error(tmp_dir):
         hypermodel=MyHyperModel(),
         max_trials=2,
         executions_per_trial=3,
-        directory=tmp_dir,
+        directory=tmp_path,
     )
 
     with pytest.raises(TypeError, match="to be a single float"):
         tuner.search()
 
 
-def test_callbacks_in_fit_kwargs(tmp_dir):
+def test_callbacks_in_fit_kwargs(tmp_path):
     tuner = keras_tuner.tuners.RandomSearch(
         build_model,
         objective="val_accuracy",
         max_trials=2,
         executions_per_trial=3,
-        directory=tmp_dir,
+        directory=tmp_path,
     )
     with patch.object(
         tuner, "_build_and_fit_model", wraps=tuner._build_and_fit_model
@@ -255,7 +321,7 @@ def test_callbacks_in_fit_kwargs(tmp_dir):
             validation_data=(VAL_INPUTS, VAL_TARGETS),
             callbacks=[
                 keras.callbacks.EarlyStopping(),
-                keras.callbacks.TensorBoard(tmp_dir),
+                keras.callbacks.TensorBoard(tmp_path),
             ],
         )
         assert len(tuner.oracle.trials) == 2
@@ -272,14 +338,14 @@ def test_callbacks_in_fit_kwargs(tmp_dir):
         ]
 
 
-def test_hypermodel_with_dynamic_space(tmp_dir):
+def test_hypermodel_with_dynamic_space(tmp_path):
     hypermodel = ExampleHyperModel()
     tuner = keras_tuner.tuners.RandomSearch(
         hypermodel,
         objective="val_accuracy",
         max_trials=2,
         executions_per_trial=3,
-        directory=tmp_dir,
+        directory=tmp_path,
     )
 
     assert tuner.hypermodel == hypermodel
@@ -299,7 +365,7 @@ def test_hypermodel_with_dynamic_space(tmp_dir):
     tuner.oracle.hyperparameters.get("shuffle")
 
 
-def test_override_compile(tmp_dir):
+def test_override_compile(tmp_path):
     class MyHyperModel(ExampleHyperModel):
         def fit(self, hp, model, *args, **kwargs):
             history = super().fit(hp, model, *args, **kwargs)
@@ -318,7 +384,7 @@ def test_override_compile(tmp_dir):
         metrics=["mse", "accuracy"],
         loss="mse",
         optimizer="rmsprop",
-        directory=tmp_dir,
+        directory=tmp_path,
     )
 
     assert tuner.oracle.objective.name == "val_mse"
@@ -338,7 +404,7 @@ def test_override_compile(tmp_dir):
     assert model.loss == "mse"
 
 
-def test_static_space(tmp_dir):
+def test_static_space(tmp_path):
     def build_model_static(hp):
         inputs = keras.Input(shape=(INPUT_DIM,))
         x = inputs
@@ -365,7 +431,7 @@ def test_static_space(tmp_dir):
         build_model_static,
         objective="val_accuracy",
         max_trials=4,
-        directory=tmp_dir,
+        directory=tmp_path,
         hyperparameters=hp,
         allow_new_entries=False,
     )
@@ -380,7 +446,7 @@ def test_static_space(tmp_dir):
     assert len(tuner.oracle.trials) == 4
 
 
-def test_static_space_errors(tmp_dir):
+def test_static_space_errors(tmp_path):
     def build_model_static(hp):
         inputs = keras.Input(shape=(INPUT_DIM,))
         x = inputs
@@ -407,7 +473,7 @@ def test_static_space_errors(tmp_dir):
             build_model_static,
             objective="val_accuracy",
             max_trials=2,
-            directory=tmp_dir,
+            directory=tmp_path,
             hyperparameters=hp,
             allow_new_entries=False,
         )
@@ -441,7 +507,7 @@ def test_static_space_errors(tmp_dir):
             build_model_static_invalid,
             objective="val_accuracy",
             max_trials=2,
-            directory=tmp_dir,
+            directory=tmp_path,
             hyperparameters=hp,
             allow_new_entries=False,
         )
@@ -453,7 +519,7 @@ def test_static_space_errors(tmp_dir):
         )
 
 
-def test_restricted_space_using_defaults(tmp_dir):
+def test_restricted_space_using_defaults(tmp_path):
     hp = keras_tuner.HyperParameters()
     hp.Int("num_layers", 1, 5, 1, default=2)
     hp.Choice("learning_rate", [0.01, 0.001, 0.0001])
@@ -462,7 +528,7 @@ def test_restricted_space_using_defaults(tmp_dir):
         build_model,
         objective="val_accuracy",
         max_trials=4,
-        directory=tmp_dir,
+        directory=tmp_path,
         hyperparameters=hp,
         allow_new_entries=True,
         tune_new_entries=False,
@@ -488,7 +554,7 @@ def test_restricted_space_using_defaults(tmp_dir):
         assert len(trial.hyperparameters.space) >= 2
 
 
-def test_restricted_space_with_custom_defaults(tmp_dir):
+def test_restricted_space_with_custom_defaults(tmp_path):
     hp = keras_tuner.HyperParameters()
     hp.Int("num_layers", 1, 3, 1, default=2)
     hp.Choice("learning_rate", [0.01, 0.001, 0.0001])
@@ -499,7 +565,7 @@ def test_restricted_space_with_custom_defaults(tmp_dir):
         build_model,
         objective="val_accuracy",
         max_trials=4,
-        directory=tmp_dir,
+        directory=tmp_path,
         hyperparameters=hp,
         allow_new_entries=True,
         tune_new_entries=False,
@@ -516,7 +582,7 @@ def test_restricted_space_with_custom_defaults(tmp_dir):
     assert len(tuner.oracle.trials) == 4
 
 
-def test_reparameterized_space(tmp_dir):
+def test_reparameterized_space(tmp_path):
     hp = keras_tuner.HyperParameters()
     hp.Int("num_layers", 1, 3, 1, default=3)
     hp.Choice("learning_rate", [0.01, 0.001, 0.0001])
@@ -526,7 +592,7 @@ def test_reparameterized_space(tmp_dir):
         seed=1337,
         objective="val_accuracy",
         max_trials=4,
-        directory=tmp_dir,
+        directory=tmp_path,
         hyperparameters=hp,
         allow_new_entries=True,
         tune_new_entries=True,
@@ -545,9 +611,9 @@ def test_reparameterized_space(tmp_dir):
     assert len(tuner.oracle.hyperparameters.space) == 5
 
 
-def test_get_best_models(tmp_dir):
+def test_get_best_models(tmp_path):
     tuner = keras_tuner.tuners.RandomSearch(
-        build_model, objective="val_accuracy", max_trials=4, directory=tmp_dir
+        build_model, objective="val_accuracy", max_trials=4, directory=tmp_path
     )
 
     tuner.search(
@@ -563,14 +629,14 @@ def test_get_best_models(tmp_dir):
     assert isinstance(models[1], keras.Model)
 
 
-def test_saving_and_reloading(tmp_dir):
+def test_saving_and_reloading(tmp_path):
 
     tuner = keras_tuner.tuners.RandomSearch(
         build_model,
         objective="val_accuracy",
         max_trials=4,
         executions_per_trial=2,
-        directory=tmp_dir,
+        directory=tmp_path,
     )
 
     tuner.search(
@@ -585,7 +651,7 @@ def test_saving_and_reloading(tmp_dir):
         objective="val_accuracy",
         max_trials=4,
         executions_per_trial=2,
-        directory=tmp_dir,
+        directory=tmp_path,
     )
     new_tuner.reload()
 
@@ -599,12 +665,12 @@ def test_saving_and_reloading(tmp_dir):
     )
 
 
-def test_subclass_model(tmp_dir):
+def test_subclass_model(tmp_path):
     tuner = keras_tuner.tuners.RandomSearch(
         build_subclass_model,
         objective="val_accuracy",
         max_trials=2,
-        directory=tmp_dir,
+        directory=tmp_path,
     )
 
     tuner.search_space_summary()
@@ -620,12 +686,12 @@ def test_subclass_model(tmp_dir):
     assert len(tuner.oracle.trials) == 2
 
 
-def test_subclass_model_loading(tmp_dir):
+def test_subclass_model_loading(tmp_path):
     tuner = keras_tuner.tuners.RandomSearch(
         build_subclass_model,
         objective="val_accuracy",
         max_trials=2,
-        directory=tmp_dir,
+        directory=tmp_path,
     )
 
     tuner.search_space_summary()
@@ -645,7 +711,7 @@ def test_subclass_model_loading(tmp_dir):
     assert best_model_score == best_trial_score
 
 
-def test_update_trial(tmp_dir):
+def test_update_trial(tmp_path):
     class MyOracle(keras_tuner.Oracle):
         def populate_space(self, _):
             values = {p.name: p.random_sample() for p in self.hyperparameters.space}
@@ -660,7 +726,7 @@ def test_update_trial(tmp_dir):
 
     my_oracle = MyOracle(objective="val_accuracy", max_trials=2)
     tuner = keras_tuner.Tuner(
-        oracle=my_oracle, hypermodel=build_model, directory=tmp_dir
+        oracle=my_oracle, hypermodel=build_model, directory=tmp_path
     )
     tuner.search(
         x=TRAIN_INPUTS,
@@ -676,7 +742,7 @@ def test_update_trial(tmp_dir):
         assert len(trial.metrics.get_history("val_accuracy")) == 1
 
 
-def test_tunable_false_hypermodel(tmp_dir):
+def test_tunable_false_hypermodel(tmp_path):
     def build_model(hp):
         input_shape = (256, 256, 3)
         inputs = tf.keras.Input(shape=input_shape)
@@ -706,7 +772,10 @@ def test_tunable_false_hypermodel(tmp_dir):
         return model
 
     tuner = keras_tuner.RandomSearch(
-        objective="val_loss", hypermodel=build_model, max_trials=4, directory=tmp_dir
+        objective="val_loss",
+        hypermodel=build_model,
+        max_trials=4,
+        directory=tmp_path,
     )
 
     x = np.random.random(size=(2, 256, 256, 3))
@@ -724,7 +793,7 @@ def test_tunable_false_hypermodel(tmp_dir):
     assert len(hps.space) == 4
 
 
-def test_get_best_hyperparameters(tmp_dir):
+def test_get_best_hyperparameters(tmp_path):
     hp1 = keras_tuner.HyperParameters()
     hp1.Fixed("a", 1)
     trial1 = keras_tuner.engine.trial.Trial(hyperparameters=hp1)
@@ -741,7 +810,7 @@ def test_get_best_hyperparameters(tmp_dir):
         objective="val_accuracy",
         hypermodel=build_model,
         max_trials=2,
-        directory=tmp_dir,
+        directory=tmp_path,
     )
 
     tuner.oracle.trials = {trial1.trial_id: trial1, trial2.trial_id: trial2}
@@ -750,8 +819,8 @@ def test_get_best_hyperparameters(tmp_dir):
     assert hps["a"] == 1
 
 
-def test_reloading_error_message(tmp_dir):
-    shared_dir = tmp_dir
+def test_reloading_error_message(tmp_path):
+    shared_dir = tmp_path
 
     tuner = keras_tuner.tuners.RandomSearch(
         build_model,
@@ -778,13 +847,13 @@ def test_reloading_error_message(tmp_dir):
         )
 
 
-def test_search_logging_verbosity(tmp_dir):
+def test_search_logging_verbosity(tmp_path):
     tuner = keras_tuner.tuners.RandomSearch(
         build_model,
         objective="val_accuracy",
         max_trials=2,
         executions_per_trial=3,
-        directory=tmp_dir,
+        directory=tmp_path,
     )
 
     with patch("sys.stdout", new=StringIO()) as output:
@@ -910,4 +979,369 @@ def test_convert_hyperparams_to_hparams():
     hparams = keras_tuner.engine.tuner_utils.convert_hyperparams_to_hparams(hps)
     _check_hparams_equal(
         hparams, {hparams_api.HParam("num_layers", hparams_api.Discrete([2])): 2}
+    )
+
+
+def test_tuning_correctness(tmp_path):
+    tuner = keras_tuner.Tuner(
+        oracle=keras_tuner.tuners.randomsearch.RandomSearchOracle(
+            objective="loss", max_trials=2, seed=1337
+        ),
+        hypermodel=MockHyperModel(),
+        directory=tmp_path,
+    )
+    tuner.search()
+    assert len(tuner.oracle.trials) == 2
+
+    m0_epochs = [float(np.average(x)) for x in MockHyperModel.mode_0]
+    m1_epochs = [float(np.average(x)) for x in MockHyperModel.mode_1]
+
+    # Score tracking correctness
+    first_trial, second_trial = sorted(
+        tuner.oracle.trials.values(), key=lambda t: t.score
+    )
+    assert first_trial.score == min(m0_epochs)
+    assert second_trial.score == min(m1_epochs)
+    assert tuner.oracle.get_best_trials(1)[0].trial_id == first_trial.trial_id
+
+
+def assert_found_best_score(tmp_path, hypermodel, tuner_class=keras_tuner.Tuner):
+    tuner = tuner_class(
+        oracle=keras_tuner.tuners.randomsearch.RandomSearchOracle(
+            objective="loss", max_trials=2, seed=1337
+        ),
+        hypermodel=hypermodel,
+        directory=tmp_path,
+    )
+    tuner.search(callbacks=[])
+    assert tuner.oracle.get_best_trials(1)[0].score == 3.0
+
+
+def test_hypermodel_fit_return_a_dict(tmp_path):
+    class MyHyperModel(MockHyperModel):
+        def fit(self, hp, model, *args, **kwargs):
+            history = super().fit(hp, model, *args, **kwargs)
+            return {
+                "loss": min(history.history["loss"]),
+                "other_metric": np.random.rand(),
+            }
+
+    assert_found_best_score(tmp_path, MyHyperModel())
+
+
+def test_hypermodel_fit_return_a_float(tmp_path):
+    class MyHyperModel(MockHyperModel):
+        def fit(self, hp, model, *args, **kwargs):
+            history = super().fit(hp, model, *args, **kwargs)
+            return min(history.history["loss"])
+
+    assert_found_best_score(tmp_path, MyHyperModel())
+
+
+def test_hypermodel_fit_return_an_int(tmp_path):
+    class MyHyperModel(MockHyperModel):
+        def fit(self, hp, model, *args, **kwargs):
+            history = super().fit(hp, model, *args, **kwargs)
+            return int(min(history.history["loss"]))
+
+    assert_found_best_score(tmp_path, MyHyperModel())
+
+
+def test_run_trial_return_none(tmp_path):
+    class MyTuner(keras_tuner.Tuner):
+        def run_trial(self, trial, *fit_args, **fit_kwargs):
+            history = self.hypermodel.build(trial.hyperparameters).fit(
+                *fit_args, **fit_kwargs
+            )
+            self.oracle.update_trial(
+                trial.trial_id, {"loss": min(history.history["loss"])}
+            )
+
+    with pytest.deprecated_call(match="float, dict, keras.callbacks.History"):
+        assert_found_best_score(tmp_path, MockHyperModel(), MyTuner)
+
+
+def test_run_trial_return_history(tmp_path):
+    class MyTuner(keras_tuner.Tuner):
+        def run_trial(self, trial, *fit_args, **fit_kwargs):
+            return self.hypermodel.build(trial.hyperparameters).fit(
+                *fit_args, **fit_kwargs
+            )
+
+    assert_found_best_score(tmp_path, MockHyperModel(), MyTuner)
+
+
+def test_run_trial_return_a_dict(tmp_path):
+    class MyTuner(keras_tuner.Tuner):
+        def run_trial(self, trial, *fit_args, **fit_kwargs):
+            history = self.hypermodel.build(trial.hyperparameters).fit(
+                *fit_args, **fit_kwargs
+            )
+            return {"loss": min(history.history["loss"])}
+
+    assert_found_best_score(tmp_path, MockHyperModel(), MyTuner)
+
+
+def test_run_trial_return_a_float(tmp_path):
+    class MyTuner(keras_tuner.Tuner):
+        def run_trial(self, trial, *fit_args, **fit_kwargs):
+            history = self.hypermodel.build(trial.hyperparameters).fit(
+                *fit_args, **fit_kwargs
+            )
+            return min(history.history["loss"])
+
+    assert_found_best_score(tmp_path, MockHyperModel(), MyTuner)
+
+
+def test_run_trial_return_float_list(tmp_path):
+    class MyTuner(keras_tuner.Tuner):
+        def run_trial(self, trial, *fit_args, **fit_kwargs):
+            ret = []
+            for i in range(3):
+                history = self.hypermodel.build(trial.hyperparameters).fit(
+                    *fit_args, **fit_kwargs
+                )
+                ret.append(min(history.history["loss"]))
+            return ret
+
+    assert_found_best_score(tmp_path, MockHyperModel(), MyTuner)
+
+
+def test_tuner_errors(tmp_path):
+    # invalid oracle
+    with pytest.raises(
+        ValueError, match="Expected `oracle` argument to be an instance of `Oracle`"
+    ):
+        tuner_module.Tuner(
+            oracle="invalid", hypermodel=build_model, directory=tmp_path
+        )
+    # invalid hypermodel
+    with pytest.raises(ValueError, match="`hypermodel` argument should be either"):
+        tuner_module.Tuner(
+            oracle=keras_tuner.tuners.randomsearch.RandomSearchOracle(
+                objective="val_accuracy", max_trials=3
+            ),
+            hypermodel="build_model",
+            directory=tmp_path,
+        )
+    # oversize model
+    with pytest.raises(RuntimeError, match="Too many consecutive oversized models"):
+        tuner = tuner_module.Tuner(
+            oracle=keras_tuner.tuners.randomsearch.RandomSearchOracle(
+                objective="val_accuracy", max_trials=3
+            ),
+            hypermodel=build_model,
+            max_model_size=4,
+            directory=tmp_path,
+        )
+        tuner.search(
+            TRAIN_INPUTS, TRAIN_TARGETS, validation_data=(VAL_INPUTS, VAL_TARGETS)
+        )
+    # TODO: test no optimizer
+
+
+@pytest.mark.skipif(
+    parse(tf.__version__) < parse("2.3.0"),
+    reason="TPUStrategy only exists in TF2.3+.",
+)
+def test_metric_direction_inferred_from_objective(tmp_path):
+    oracle = keras_tuner.tuners.randomsearch.RandomSearchOracle(
+        objective=keras_tuner.Objective("a", "max"), max_trials=1
+    )
+    oracle._set_project_dir(tmp_path, "untitled_project")
+    trial = oracle.create_trial("tuner0")
+    oracle.update_trial(trial.trial_id, {"a": 1})
+    trial = oracle.get_trial(trial.trial_id)
+    assert trial.metrics.get_direction("a") == "max"
+
+    oracle = keras_tuner.tuners.randomsearch.RandomSearchOracle(
+        objective=keras_tuner.Objective("a", "min"), max_trials=1
+    )
+    oracle._set_project_dir(tmp_path, "untitled_project2")
+    trial = oracle.create_trial("tuner0")
+    oracle.update_trial(trial.trial_id, {"a": 1})
+    trial = oracle.get_trial(trial.trial_id)
+    assert trial.metrics.get_direction("a") == "min"
+
+
+def test_overwrite_true(tmp_path):
+    tuner = keras_tuner.tuners.RandomSearch(
+        hypermodel=build_model,
+        objective="val_accuracy",
+        max_trials=2,
+        directory=tmp_path,
+    )
+    tuner.search(
+        TRAIN_INPUTS, TRAIN_TARGETS, validation_data=(VAL_INPUTS, VAL_TARGETS)
+    )
+    assert len(tuner.oracle.trials) == 2
+
+    new_tuner = keras_tuner.tuners.RandomSearch(
+        hypermodel=build_model,
+        objective="val_accuracy",
+        max_trials=2,
+        directory=tmp_path,
+        overwrite=True,
+    )
+    assert len(new_tuner.oracle.trials) == 0
+
+
+def test_correct_display_trial_number(tmp_path):
+    tuner = keras_tuner.tuners.RandomSearch(
+        hypermodel=build_model,
+        objective="val_accuracy",
+        max_trials=2,
+        directory=tmp_path,
+    )
+    tuner.search(
+        TRAIN_INPUTS, TRAIN_TARGETS, validation_data=(VAL_INPUTS, VAL_TARGETS)
+    )
+    new_tuner = keras_tuner.tuners.RandomSearch(
+        hypermodel=build_model,
+        objective="val_accuracy",
+        max_trials=6,
+        directory=tmp_path,
+        overwrite=False,
+    )
+    new_tuner.search(
+        TRAIN_INPUTS, TRAIN_TARGETS, validation_data=(VAL_INPUTS, VAL_TARGETS)
+    )
+    assert len(new_tuner.oracle.trials) == new_tuner._display.trial_number
+
+
+def test_error_on_unknown_objective_direction(tmp_path):
+    with pytest.raises(ValueError, match="Could not infer optimization direction"):
+        keras_tuner.tuners.RandomSearch(
+            hypermodel=build_model,
+            objective="custom_metric",
+            max_trials=2,
+            directory=tmp_path,
+        )
+
+
+def test_callbacks_run_each_execution(tmp_path):
+    callback_instances = set()
+
+    class LoggingCallback(keras.callbacks.Callback):
+        def on_train_begin(self, logs):
+            callback_instances.add(id(self))
+
+    logging_callback = LoggingCallback()
+    tuner = keras_tuner.tuners.RandomSearch(
+        hypermodel=build_model,
+        objective="val_accuracy",
+        max_trials=2,
+        executions_per_trial=3,
+        directory=tmp_path,
+    )
+    tuner.search(
+        TRAIN_INPUTS,
+        TRAIN_TARGETS,
+        validation_data=(VAL_INPUTS, VAL_TARGETS),
+        callbacks=[logging_callback],
+    )
+
+    assert len(callback_instances) == 6
+
+
+def test_build_and_fit_model(tmp_path):
+    class MyTuner(keras_tuner.tuners.RandomSearch):
+        def _build_and_fit_model(self, trial, *args, **kwargs):
+            self.was_called = True
+            return super()._build_and_fit_model(trial, *args, **kwargs)
+
+    tuner = MyTuner(
+        hypermodel=build_model,
+        objective="val_accuracy",
+        max_trials=2,
+        executions_per_trial=3,
+        directory=tmp_path,
+    )
+
+    tuner.run_trial(
+        tuner.oracle.create_trial("tuner0"),
+        TRAIN_INPUTS,
+        TRAIN_TARGETS,
+        validation_data=(VAL_INPUTS, VAL_TARGETS),
+    )
+
+    assert tuner.was_called
+
+
+def test_build_and_fit_model_in_tuner(tmp_path):
+    class MyTuner(tuner_module.Tuner):
+        def _build_and_fit_model(self, trial, *args, **kwargs):
+            self.was_called = True
+            return super()._build_and_fit_model(trial, *args, **kwargs)
+
+    tuner = MyTuner(
+        oracle=keras_tuner.tuners.randomsearch.RandomSearchOracle(
+            objective="val_loss",
+            max_trials=2,
+        ),
+        hypermodel=build_model,
+        directory=tmp_path,
+    )
+
+    tuner.run_trial(
+        tuner.oracle.create_trial("tuner0"),
+        TRAIN_INPUTS,
+        TRAIN_TARGETS,
+        validation_data=(VAL_INPUTS, VAL_TARGETS),
+    )
+
+    assert tuner.was_called
+
+
+def test_init_build_all_hps_in_all_conditions(tmp_path):
+    class ConditionalHyperModel(MockHyperModel):
+        def build(self, hp):
+            model_type = hp.Choice("model_type", ["cnn", "mlp"])
+            with hp.conditional_scope("model_type", ["cnn"]):
+                if model_type == "cnn":
+                    sub_cnn = hp.Choice("sub_cnn", ["a", "b"])
+                    with hp.conditional_scope("sub_cnn", ["a"]):
+                        if sub_cnn == "a":
+                            hp.Int("n_filters_a", 2, 4)
+                    with hp.conditional_scope("sub_cnn", ["b"]):
+                        if sub_cnn == "b":
+                            hp.Int("n_filters_b", 6, 8)
+            with hp.conditional_scope("model_type", ["mlp"]):
+                if model_type == "mlp":
+                    sub_mlp = hp.Choice("sub_mlp", ["a", "b"])
+                    with hp.conditional_scope("sub_mlp", ["a"]):
+                        if sub_mlp == "a":
+                            hp.Int("n_units_a", 2, 4)
+                    with hp.conditional_scope("sub_mlp", ["b"]):
+                        if sub_mlp == "b":
+                            hp.Int("n_units_b", 6, 8)
+            more_block = hp.Boolean("more_block", default=False)
+            with hp.conditional_scope("more_block", [True]):
+                if more_block:
+                    hp.Int("new_block_hp", 1, 3)
+            return super().build(hp)
+
+    def name_in_hp(name, hp):
+        return any([name == single_hp.name for single_hp in hp.space])
+
+    class MyTuner(tuner_module.Tuner):
+        def _populate_initial_space(self):
+            super()._populate_initial_space()
+            hp = self.oracle.hyperparameters
+            assert name_in_hp("model_type", hp)
+            assert name_in_hp("sub_cnn", hp)
+            assert name_in_hp("n_filters_a", hp)
+            assert name_in_hp("n_filters_b", hp)
+            assert name_in_hp("sub_mlp", hp)
+            assert name_in_hp("n_units_a", hp)
+            assert name_in_hp("n_units_b", hp)
+            assert name_in_hp("more_block", hp)
+            assert name_in_hp("new_block_hp", hp)
+
+    MyTuner(
+        oracle=keras_tuner.tuners.randomsearch.RandomSearchOracle(
+            objective="loss", max_trials=2, seed=1337
+        ),
+        hypermodel=ConditionalHyperModel(),
+        directory=tmp_path,
     )
