+++ setup.cfg
@@ -26,6 +26,7 @@ exclude_lines =
     pragma: no cover
     @abstract
     raise NotImplementedError
+omit = *test*
 
 [flake8]
