+++ setup.py
@@ -59,5 +59,5 @@ setup(
         "Topic :: Scientific/Engineering",
         "Topic :: Software Development",
     ],
-    packages=find_packages(exclude=("tests",)),
+    packages=find_packages(exclude=("*test*",)),
 )
